+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/UTxO.hs
@@ -26,11 +26,18 @@ module Cardano.Ledger.Alonzo.UTxO (
   getRewardingScriptsNeeded,
   getMintingScriptsNeeded,
   getAlonzoScriptsHashesNeeded,
+  scriptsNeededAlonzoStAnnTx,
   zipAsIxItem,
 
   -- * Scripts provided
   scriptsProvidedAlonzoStAnnTx,
   resolveNeededPlutusScriptsWithPurpose,
+
+  -- * Plutus scripts with context
+  plutusScriptsWithContextAlonzoStAnnTx,
+
+  -- * Plutus languages used
+  plutusLanguagesUsedAlonzoStAnnTx,
 
   -- * Datums needed
   getInputDataHashesTxBody,
@@ -42,6 +49,7 @@ module Cardano.Ledger.Alonzo.UTxO (
 import Cardano.Ledger.Address (accountAddressCredentialL)
 import Cardano.Ledger.Alonzo.Core
 import Cardano.Ledger.Alonzo.Era (AlonzoEra)
+import Cardano.Ledger.Alonzo.Plutus.Context (CollectError)
 import Cardano.Ledger.Alonzo.Scripts (lookupPlutusScript, plutusScriptLanguage)
 import Cardano.Ledger.Alonzo.State ()
 import Cardano.Ledger.Alonzo.Tx (AlonzoStAnnTx (..))
@@ -51,7 +59,7 @@ import Cardano.Ledger.Credential (credScriptHash)
 import Cardano.Ledger.Keys (asWitness)
 import Cardano.Ledger.Mary.UTxO (getConsumedMaryValue, getProducedMaryValue)
 import Cardano.Ledger.Mary.Value (PolicyID (..))
-import Cardano.Ledger.Plutus (Language (..))
+import Cardano.Ledger.Plutus (Language (..), PlutusWithContext)
 import Cardano.Ledger.Plutus.Data (Data, Datum (..))
 import Cardano.Ledger.Shelley.UTxO (
   getShelleyMinFeeTxUtxo,
@@ -68,6 +76,7 @@ import Cardano.Ledger.State (
 import Cardano.Ledger.TxIn
 import Control.DeepSeq (NFData)
 import Data.Foldable as F (foldl', toList)
+import Data.List.NonEmpty (NonEmpty)
 import qualified Data.Map.Strict as Map
 import Data.Maybe (catMaybes, fromMaybe, isJust)
 import qualified Data.Set as Set
@@ -134,12 +143,26 @@ class EraUTxO era => AlonzoEraUTxO era where
 
   scriptsProvidedStAnnTx :: StAnnTx l era -> ScriptsProvided era
 
+  scriptsNeededStAnnTx :: StAnnTx l era -> ScriptsNeeded era
+
+  plutusScriptsWithContextStAnnTx ::
+    StAnnTx l era ->
+    Either (NonEmpty (CollectError era)) [PlutusWithContext]
+
+  plutusLanguagesUsedStAnnTx :: StAnnTx l era -> Set.Set Language
+
 instance AlonzoEraUTxO AlonzoEra where
   getSupplementalDataHashes _ = getAlonzoSupplementalDataHashes
 
   getSpendingDatum = getAlonzoSpendingDatum
 
   scriptsProvidedStAnnTx = scriptsProvidedAlonzoStAnnTx
+
+  scriptsNeededStAnnTx = scriptsNeededAlonzoStAnnTx
+
+  plutusScriptsWithContextStAnnTx = plutusScriptsWithContextAlonzoStAnnTx
+
+  plutusLanguagesUsedStAnnTx = plutusLanguagesUsedAlonzoStAnnTx
 
 scriptsProvidedAlonzoStAnnTx ::
   ( EraTxLevel era
@@ -149,6 +172,36 @@ scriptsProvidedAlonzoStAnnTx ::
   AlonzoStAnnTx l era -> ScriptsProvided era
 scriptsProvidedAlonzoStAnnTx stAnnTx =
   withTopTxLevelOnly stAnnTx $ \AlonzoStAnnTx {asatScriptsProvided} -> asatScriptsProvided
+
+scriptsNeededAlonzoStAnnTx ::
+  ( EraTxLevel era
+  , STxLevel l era ~ STxTopLevel l era
+  , STxLevel TopTx era ~ STxTopLevel TopTx era
+  ) =>
+  AlonzoStAnnTx l era -> ScriptsNeeded era
+scriptsNeededAlonzoStAnnTx stAnnTx =
+  withTopTxLevelOnly stAnnTx $ \AlonzoStAnnTx {asatScriptsNeeded} -> asatScriptsNeeded
+
+plutusScriptsWithContextAlonzoStAnnTx ::
+  ( EraTxLevel era
+  , STxLevel l era ~ STxTopLevel l era
+  , STxLevel TopTx era ~ STxTopLevel TopTx era
+  ) =>
+  AlonzoStAnnTx l era ->
+  Either (NonEmpty (CollectError era)) [PlutusWithContext]
+plutusScriptsWithContextAlonzoStAnnTx stAnnTx =
+  withTopTxLevelOnly stAnnTx $
+    \AlonzoStAnnTx {asatPlutusScriptsWithContext} -> asatPlutusScriptsWithContext
+
+plutusLanguagesUsedAlonzoStAnnTx ::
+  ( EraTxLevel era
+  , STxLevel l era ~ STxTopLevel l era
+  , STxLevel TopTx era ~ STxTopLevel TopTx era
+  ) =>
+  AlonzoStAnnTx l era -> Set.Set Language
+plutusLanguagesUsedAlonzoStAnnTx stAnnTx =
+  withTopTxLevelOnly stAnnTx $
+    \AlonzoStAnnTx {asatPlutusLanguagesUsed} -> asatPlutusLanguagesUsed
 
 getAlonzoSupplementalDataHashes ::
   (EraTxBody era, AlonzoEraTxOut era) =>