+++ eras/babbage/impl/src/Cardano/Ledger/Babbage/Rules/Utxos.hs
@@ -19,10 +19,7 @@ module Cardano.Ledger.Babbage.Rules.Utxos (
 ) where
 
 import Cardano.Ledger.Alonzo.Plutus.Context (EraPlutusContext)
-import Cardano.Ledger.Alonzo.Plutus.Evaluate (
-  collectPlutusScriptsWithContext,
-  evalPlutusScripts,
- )
+import Cardano.Ledger.Alonzo.Plutus.Evaluate (evalPlutusScripts)
 import qualified Cardano.Ledger.Alonzo.Rules as Alonzo
 import Cardano.Ledger.Alonzo.UTxO (
   AlonzoEraUTxO (..),
@@ -32,12 +29,7 @@ import Cardano.Ledger.Babbage.Core
 import Cardano.Ledger.Babbage.Era (BabbageEra, BabbageUTXOS)
 import Cardano.Ledger.Babbage.Rules.Ppup ()
 import Cardano.Ledger.Babbage.State
-import Cardano.Ledger.BaseTypes (
-  ShelleyBase,
-  StrictMaybe,
-  epochInfo,
-  systemStart,
- )
+import Cardano.Ledger.BaseTypes (ShelleyBase, StrictMaybe)
 import Cardano.Ledger.Binary (EncCBOR (..))
 import Cardano.Ledger.Plutus.Evaluate (
   ScriptFailure (..),
@@ -46,7 +38,6 @@ import Cardano.Ledger.Plutus.Evaluate (
 import Cardano.Ledger.Shelley.PParams (Update)
 import qualified Cardano.Ledger.Shelley.Rules as Shelley
 import Control.Monad (forM_)
-import Control.Monad.Trans.Reader (asks)
 import Control.State.Transition.Extended
 import Data.List.NonEmpty (nonEmpty)
 import qualified Debug.Trace as Debug
@@ -114,33 +105,24 @@ utxosTransition ::
   forall era.
   ( AlonzoEraTx era
   , ShelleyEraTxBody era
-  , BabbageEraTxBody era
   , AlonzoEraUTxO era
-  , ScriptsNeeded era ~ AlonzoScriptsNeeded era
   , EraCertState era
-  , EraStake era
-  , EraGov era
-  , GovState era ~ ShelleyGovState era
   , Environment (EraRule "PPUP" era) ~ Shelley.PpupEnv era
   , Signal (EraRule "PPUP" era) ~ StrictMaybe (Update era)
   , Embed (EraRule "PPUP" era) (BabbageUTXOS era)
   , State (EraRule "PPUP" era) ~ ShelleyGovState era
-  , EncCBOR (EraRuleFailure "PPUP" era)
-  , Eq (EraRuleFailure "PPUP" era)
-  , Show (EraRuleFailure "PPUP" era)
-  , EraPlutusContext era
   , EraRule "UTXOS" era ~ BabbageUTXOS era
   , InjectRuleFailure "UTXOS" Alonzo.AlonzoUtxosPredFailure era
   , InjectRuleEvent "UTXOS" Alonzo.AlonzoUtxosEvent era
   ) =>
   TransitionRule (BabbageUTXOS era)
 utxosTransition =
-  judgmentContext >>= \(TRC (Alonzo.UtxosEnv _ pp _ utxo, pup, stAnnTx)) -> do
+  judgmentContext >>= \(TRC (_, pup, stAnnTx)) -> do
     let tx = stAnnTx ^. txStAnnTxG
     case tx ^. isValidTxL of
       IsValid True -> babbageEvalScriptsTxValid
       IsValid False -> do
-        babbageEvalScriptsTxInvalid @era pp tx utxo
+        babbageEvalScriptsTxInvalid @era stAnnTx
         pure pup
 
 -- ===================================================================
@@ -148,25 +130,16 @@ utxosTransition =
 expectScriptsToPass ::
   forall era.
   ( AlonzoEraTx era
-  , EraPlutusContext era
   , AlonzoEraUTxO era
-  , ScriptsNeeded era ~ AlonzoScriptsNeeded era
-  , STS (EraRule "UTXOS" era)
   , InjectRuleFailure "UTXOS" Alonzo.AlonzoUtxosPredFailure era
-  , BaseM (EraRule "UTXOS" era) ~ ShelleyBase
   , InjectRuleEvent "UTXOS" Alonzo.AlonzoUtxosEvent era
   ) =>
-  PParams era ->
-  Tx TopTx era ->
-  UTxO era ->
+  StAnnTx TopTx era ->
   Rule (EraRule "UTXOS" era) 'Transition ()
-expectScriptsToPass pp tx utxo = do
-  sysSt <- liftSTS $ asks systemStart
-  ei <- liftSTS $ asks epochInfo
+expectScriptsToPass stAnnTx = do
+  let tx = stAnnTx ^. txStAnnTxG
   {- sLst := collectTwoPhaseScriptInputs pp tx utxo -}
-  let
-    scriptsWithContextEither =
-      collectPlutusScriptsWithContext ei sysSt pp tx utxo
+  let scriptsWithContextEither = plutusScriptsWithContextStAnnTx stAnnTx
   (() <$ scriptsWithContextEither) ?!: (injectFailure . Alonzo.CollectErrors)
   {- isValid tx = evalScripts tx sLst = True -}
   Alonzo.when2Phase $
@@ -187,20 +160,17 @@ babbageEvalScriptsTxValid ::
   , AlonzoEraUTxO era
   , ShelleyEraTxBody era
   , EraCertState era
-  , ScriptsNeeded era ~ AlonzoScriptsNeeded era
-  , STS (BabbageUTXOS era)
   , Environment (EraRule "PPUP" era) ~ Shelley.PpupEnv era
   , Signal (EraRule "PPUP" era) ~ StrictMaybe (Update era)
   , Embed (EraRule "PPUP" era) (BabbageUTXOS era)
   , State (EraRule "PPUP" era) ~ ShelleyGovState era
-  , EraPlutusContext era
   , InjectRuleFailure "UTXOS" Alonzo.AlonzoUtxosPredFailure era
   , EraRule "UTXOS" era ~ BabbageUTXOS era
   , InjectRuleEvent "UTXOS" Alonzo.AlonzoUtxosEvent era
   ) =>
   TransitionRule (BabbageUTXOS era)
 babbageEvalScriptsTxValid = do
-  TRC (Alonzo.UtxosEnv slot pp certState utxo, pup, stAnnTx) <-
+  TRC (Alonzo.UtxosEnv slot pp certState _utxo, pup, stAnnTx) <-
     judgmentContext
   let tx = stAnnTx ^. txStAnnTxG
       txBody = tx ^. bodyTxL
@@ -214,7 +184,7 @@ babbageEvalScriptsTxValid = do
       TRC (Shelley.PPUPEnv slot pp genDelegs, pup, txBody ^. updateTxBodyL)
 
   () <- pure $! Debug.traceEvent Alonzo.validBegin ()
-  expectScriptsToPass pp tx utxo
+  expectScriptsToPass stAnnTx
   () <- pure $! Debug.traceEvent Alonzo.validEnd ()
 
   pure updatedGovState
@@ -222,26 +192,17 @@ babbageEvalScriptsTxValid = do
 babbageEvalScriptsTxInvalid ::
   forall era.
   ( AlonzoEraTx era
-  , EraPlutusContext era
   , AlonzoEraUTxO era
-  , ScriptsNeeded era ~ AlonzoScriptsNeeded era
   , InjectRuleFailure "UTXOS" Alonzo.AlonzoUtxosPredFailure era
   , InjectRuleEvent "UTXOS" Alonzo.AlonzoUtxosEvent era
-  , BaseM (EraRule "UTXOS" era) ~ ShelleyBase
-  , STS (EraRule "UTXOS" era)
   ) =>
-  PParams era ->
-  Tx TopTx era ->
-  UTxO era ->
+  StAnnTx TopTx era ->
   Rule (EraRule "UTXOS" era) 'Transition ()
-babbageEvalScriptsTxInvalid pp tx utxo = do
-  sysSt <- liftSTS $ asks systemStart
-  ei <- liftSTS $ asks epochInfo
+babbageEvalScriptsTxInvalid stAnnTx = do
+  let tx = stAnnTx ^. txStAnnTxG
   () <- pure $! Debug.traceEvent Alonzo.invalidBegin ()
   {- sLst := collectTwoPhaseScriptInputs pp tx utxo -}
-  let
-    scriptsWithContextEither =
-      collectPlutusScriptsWithContext ei sysSt pp tx utxo
+  let scriptsWithContextEither = plutusScriptsWithContextStAnnTx stAnnTx
   (() <$ scriptsWithContextEither) ?!: (injectFailure . Alonzo.CollectErrors)
   {- isValid tx = evalScripts tx sLst = False -}
   Alonzo.when2Phase $